# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title: VIB DevOps Ecosystem Platform — Combined API
  version: "2.0.0"
  description: |
    API contract for 4 modules added in BRD-001 v1.1 (FR-019 – FR-032).
    Auth: Bearer JWT (ADR-003). All write operations produce audit log (NFR-007).
    Banking-grade: all endpoints except /health require Authorization header.
  contact:
    name: Backend Dev (Agent 06)
servers:
  - url: http://localhost:8001
    description: PPG System — Annual Plans, Project Objects, Reports
  - url: http://localhost:8002
    description: BA Workflow — BA Documents v2
  - url: http://localhost:8003
    description: Test Platform — Test Documents
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/reports/connections:
    get:
      summary: Cross-project connection report (FR-026)
      operationId: crossProjectConnections
      tags: [reports]
      parameters:
        - name: object_name
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
            format: uuid
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: direction
          in: query
          schema:
            type: string
            enum: [in, out, both]
            default: both
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Connection report with inbound and outbound connections
        "404":
          description: No objects found
        "422":
          description: Neither object_name nor object_id provided
  /api/v1/reports/annual-plan-summary/{plan_id}:
    get:
      summary: Annual plan dashboard summary (FR-022)
      operationId: annualPlanSummary
      tags: [reports]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Dashboard with dod_completion_pct, projects_by_status, project details

  # =========================================================================
  # Module 3 — BA Documents v2 (BA Workflow :8002)
  # =========================================================================
...